Checking argument for null in actions

+++ OrangeJuice.Server.Api/Controllers/RatingController.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using System.Web.Http;
 
 using OrangeJuice.Server.Api.Models;
@@ -23,6 +24,9 @@ namespace OrangeJuice.Server.Api.Controllers
 		#region HTTP methods
 		public async Task<IHttpActionResult> GetRating([FromUri]RatingSearchCriteria searchCriteria)
 		{
+			if (searchCriteria == null)
+				throw new ArgumentNullException();
+
 			IRating rating = await _ratingRepository.Search(searchCriteria.UserId, searchCriteria.ProductId);
 			if (rating == null)
 				return NotFound();
@@ -33,6 +37,9 @@ namespace OrangeJuice.Server.Api.Controllers
 		[Route("api/product/{productId}/rating")]
 		public async Task<IHttpActionResult> GetRatings([FromUri]RatingsSearchCriteria searchCriteria)
 		{
+			if (searchCriteria == null)
+				throw new ArgumentNullException();
+
 			var ratings = await _ratingRepository.SearchAll(searchCriteria.ProductId);
 			if (ratings == null)
 				return NotFound();
@@ -42,6 +49,9 @@ namespace OrangeJuice.Server.Api.Controllers
 
 		public async Task<IHttpActionResult> PostRating(RatingModel ratingModel)
 		{
+			if (ratingModel == null)
+				throw new ArgumentNullException();
+
 			await _ratingRepository.AddOrUpdate(ratingModel.UserId, ratingModel.ProductId, ratingModel.Value, ratingModel.Comment);
 
 			return Ok();
@@ -49,6 +59,9 @@ namespace OrangeJuice.Server.Api.Controllers
 
 		public async Task<IHttpActionResult> DeleteRating([FromUri]RatingSearchCriteria searchCriteria)
 		{
+			if (searchCriteria == null)
+				throw new ArgumentNullException();
+
 			await _ratingRepository.Delete(searchCriteria.UserId, searchCriteria.ProductId);
 
 			return Ok();
